Reorganized exception modules
