+++ symphony/bdk/core/config/exception.py
@@ -1,3 +1,12 @@
+"""Module containing all configuration related exception.
+"""
+
+
+class BdkConfigException(Exception):
+    """Exception class raised when configuration is invalid.
+    """
+
+
 class BotNotConfiguredException(Exception):
     """Thrown when the bot configuration is not specified."""
