--- symphony/bdk/core/config/exception/bdk_config_exception.py
@@ -1,3 +1,0 @@
-class BdkConfigException(Exception):
-    """Exception class raised when configuration is invalid.
-    """
