+++ tests/core/config/test_bdk_config_parser.py
@@ -1,4 +1,4 @@
-from symphony.bdk.core.config.exception.bdk_config_exception import BdkConfigException
+from symphony.bdk.core.config.exception import BdkConfigException
 from symphony.bdk.core.config.bdk_config_parser import BdkConfigParser
 
 from tests.utils.resource_utils import get_config_resource_filepath
@@ -7,7 +7,7 @@ import pytest
 
 @pytest.fixture(params=["invalid_config.yaml"])
 def invalid_config_path(request):
-    return get_config_resource_filepath(request.param, as_text=False)
+    return get_config_resource_filepath(request.param)
 
 
 def test_parse_config_json():
